[#100] 403 코드가 들어오면 자동으로 로그아웃 되게끔 수정

+++ Shoak/Common/AccountManager.swift
@@ -68,10 +68,12 @@ class AccountManager: @unchecked Sendable {
 
     public func signOut() async -> Result<Void, NetworkError> {
         let response = await userUseCase.signOut()
-        tokenUseCase.deleteAllTokens()
+        if case .success = response {
+            tokenUseCase.deleteAllTokens()
 #if os(iOS)
-        await UIApplication.shared.unregisterForRemoteNotifications()
+            await UIApplication.shared.unregisterForRemoteNotifications()
 #endif
+        }
         return response
     }
